Working UI!

+++ Taj/PalaceConnection.cs
@@ -160,7 +160,7 @@ namespace Taj
                                             Palace.HTTPServer = msg_httpsv.Location;
                                             break;
                                         default:
-                                            Debug.WriteLine("Unknown EvT: 0x{0:X8}", msg.eventType);
+                                            Debug.WriteLine("Unknown EvT: 0x{0:X8}", (uint)msg.eventType);
                                             Reader.Read(new byte[msg.length], 0, msg.length);
                                             break;
                                     }
